Move page size to separate interface
