+++ src/Viewer.tsx
@@ -14,7 +14,8 @@ import defaultLayout from './layouts/defaultLayout';
 import defaultToolbar from './layouts/defaultToolbar';
 import Inner from './layouts/Inner';
 import { Layout } from './layouts/Layout';
-import PageSizeCalculator, { PageSize } from './layouts/PageSizeCalculator';
+import PageSize from './layouts/PageSize'; 
+import PageSizeCalculator from './layouts/PageSizeCalculator';
 import { RenderPage } from './layouts/RenderPage';
 import { RenderToolbar } from './layouts/ToolbarSlot';
 import DocumentLoader, { RenderError } from './loader/DocumentLoader';
