+++ src/layouts/PageSize.ts
@@ -1,0 +1,15 @@
+/**
+ * A React component to view a PDF document
+ *
+ * @see https://react-pdf-viewer.dev
+ * @license https://react-pdf-viewer.dev/license
+ * @copyright 2019-2020 Nguyen Huu Phuoc <me@phuoc.ng>
+ */
+
+interface PageSize {
+    pageHeight: number;
+    pageWidth: number;
+    scale: number;
+}
+
+export default PageSize;
