+++ src/layouts/PageSizeCalculator.tsx
@@ -13,22 +13,17 @@ import ThemeContext from '../theme/ThemeContext';
 import PdfJs from '../vendors/PdfJs';
 import { decrease } from '../zoom/zoomingLevel';
 import './pageSizeCalculator.less';
-
-interface PageSizeProps {
-    pageHeight: number;
-    pageWidth: number;
-    scale: number;
-}
+import PageSize from './PageSize';
 
 interface PageSizeCalculatorProps {
     doc: PdfJs.PdfDocument;
-    render(pageSize: PageSizeProps): React.ReactElement;
+    render(pageSize: PageSize): React.ReactElement;
 }
 
 const PageSizeCalculator: React.FC<PageSizeCalculatorProps> = ({ doc, render }) => {
     const theme = React.useContext(ThemeContext);
     const pagesRef = React.useRef<HTMLDivElement | null>(null);
-    const [pageSize, setPageSize] = React.useState<PageSizeProps>({
+    const [pageSize, setPageSize] = React.useState<PageSize>({
         pageHeight: 0,
         pageWidth: 0,
         scale: 1,
@@ -37,8 +32,10 @@ const PageSizeCalculator: React.FC<PageSizeCalculatorProps> = ({ doc, render }) 
     React.useEffect(() => {
         doc.getPage(1).then((pdfPage) => {
             const viewport = pdfPage.getViewport({ scale: 1 });
+
             const w = viewport.width;
             const h = viewport.height;
+            console.log(viewport);
 
             const pagesEle = pagesRef.current;
             if (!pagesEle) {
@@ -70,4 +67,3 @@ const PageSizeCalculator: React.FC<PageSizeCalculatorProps> = ({ doc, render }) 
 };
 
 export default PageSizeCalculator;
-export type PageSize = PageSizeProps;
